requirements.txt
gradio
google-generativeai
spacy
yake